Refactor: move utilites to the separate directory
