--- Tasks/MavenV3/utils.ts
@@ -1,125 +1,0 @@
-import * as path from 'path';
-import * as fs from 'fs';
-import * as tl from 'azure-pipelines-task-lib/task';
-import * as xml2js from 'xml2js';
-import * as shelljs from 'shelljs'
-
-export async function readFile(filePath: string, encoding: string) {
-
-    tl.debug('Reading file at path ' + filePath)
-
-    return new Promise<string>((resolve, reject) =>
-        fs.readFile(filePath, (err, buffer) => {
-            const fileData = buffer.toString(encoding)
-            resolve(fileData)
-        })
-    )
-}
-
-function fixEncoding(str: string) {
-    if (str.charCodeAt(0) === 0xFEFF) {
-        return str.slice(1);
-    }
-    return str;
-}
-
-export async function convertXmlStringToJson(xmlContent: string) {
-    tl.debug("Converting XML file to JSON");
-    const fixedXml = fixEncoding(xmlContent)
-    try {
-        const json = await xml2js.parseStringPromise(fixedXml)
-
-        return json;
-    }
-    catch (err) {
-        tl.error("Error when conveting the xml to json: " + err)
-        throw err
-    }
-}
-
-export async function readXmlFileAsJson(filePath: string): Promise<any> {
-    try {
-        const xml = await readFile(filePath, "utf-8")
-        const json = await convertXmlStringToJson(xml)
-
-        return json
-    }
-    catch (err) {
-        tl.error(`Error when reading xml file: ${err}`)
-        throw err
-    }
-}
-
-
-export function writeJsonAsXmlFile(filePath: string, jsonContent: any) {
-    try {
-        const builder = new xml2js.Builder();
-        tl.debug("Writing JSON as XML file: " + filePath);
-
-        const xml = builder.buildObject(jsonContent).replace(/&#xD;/g, "");
-        writeFile(filePath, xml);
-    }
-    catch (err) {
-        tl.error(`Error when writing the json to the xml file: ${err}`)
-        throw err
-    }
-}
-
-export function writeFile(filePath: string, fileContent: string) {
-    try {
-        const dirname = path.dirname(filePath)
-        if (!fs.existsSync(dirname)) {
-            fs.mkdirSync(dirname);
-        }
-
-        fs.writeFileSync(filePath, fileContent, { encoding: "utf-8" });
-    }
-    catch (err) {
-        tl.error(`Error when writing to the file:${err}`)
-        throw err
-    }
-}
-
-export function addPropToJson(obj: any, propName: string, value: any): void {
-    tl.debug("Adding property to JSON: " + propName);
-    if (typeof obj === "undefined") {
-        obj = {};
-    }
-
-    if (obj instanceof Array) {
-        const propNode = obj.find(o => o[propName]);
-        if (propNode) {
-            obj = propNode;
-        }
-    }
-
-    if (propName in obj) {
-        if (obj[propName] instanceof Array) {
-            obj[propName].push(value);
-        } else if (typeof obj[propName] !== "object") {
-            obj[propName] = [obj[propName], value];
-        }
-    } else if (obj instanceof Array) {
-        const prop = {};
-        prop[propName] = value;
-        obj.push(prop);
-    } else {
-        obj[propName] = value;
-    }
-}
-
-export function copyFile(sourcePath: string, destinationPath: string): void {
-    shelljs.cp('-f', sourcePath, destinationPath);
-    checkShell('cp', false);
-}
-
-function checkShell(cmd: string, continueOnError?: boolean): void {
-    const shellError: string = shelljs.error();
-    if (shellError) {
-        tl.debug(cmd + ' failed');
-
-        if (!continueOnError) {
-            throw new Error(shellError);
-        }
-    }
-}
